Check for odd rows/columns in build_monotonic
When this operator is built for an image with an even number of
rows and/or columns, the result is a catastrophic failure.
This commit adds a check that the user has already projected
the image to a space with an odd number of rows and columns.

It also changes the name of the proximal operator, which
will be useful in the future when implementing a version
of monotonicity that doesn't just look at the radial
direction, but at all directions coming from the peak.
This future change may help model spiral galaxies
more accurately.